architecture: better audit remediation
