+++ server/services/memory.js
@@ -9,7 +9,7 @@ import { writeFile, readdir, rm } from 'fs/promises';
 import { existsSync } from 'fs';
 import { join } from 'path';
 import { v4 as uuidv4 } from '../lib/uuid.js';
-import { cosEvents } from './cos.js';
+import { cosEvents } from './cosEvents.js';
 import { findTopK, findAboveThreshold, clusterBySimilarity } from '../lib/vectorMath.js';
 import * as notifications from './notifications.js';
 import { ensureDir, ensureDirs, readJSONFile, PATHS } from '../lib/fileUtils.js';
