+++ server/services/memory.test.js
@@ -18,7 +18,10 @@ vi.mock('../lib/uuid.js', () => ({
   v4: vi.fn().mockReturnValue('test-uuid-1234')
 }));
 
-// Mock cosEvents
+// Mock cosEvents (memory.js imports from cosEvents.js directly)
+vi.mock('./cosEvents.js', () => ({
+  cosEvents: new EventEmitter()
+}));
 vi.mock('./cos.js', () => ({
   cosEvents: new EventEmitter()
 }));
@@ -79,7 +82,7 @@ import * as memoryBM25 from './memoryBM25.js';
 import * as notifications from './notifications.js';
 import { findTopK, findAboveThreshold, clusterBySimilarity } from '../lib/vectorMath.js';
 import { generateSummary, decrementAgentPendingApproval } from './memoryConfig.js';
-import { cosEvents } from './cos.js';
+import { cosEvents } from './cosEvents.js';
 import {
   createMemory,
   getMemory,
