+++ server/services/socket.js
@@ -61,14 +61,20 @@ export function initSocket(io) {
 
     // Handle streaming app detection
     socket.on('detect:start', async (rawData) => {
-      const data = validateSocketData(detectStartSchema, rawData, socket, 'detect:start');
-      if (!data) return;
-      console.log(`🔍 Starting detection: ${data.path}`);
-      await streamDetection(socket, data.path);
+      try {
+        const data = validateSocketData(detectStartSchema, rawData, socket, 'detect:start');
+        if (!data) return;
+        console.log(`🔍 Starting detection: ${data.path}`);
+        await streamDetection(socket, data.path);
+      } catch (err) {
+        console.error(`❌ Socket handler error [detect:start]: ${err.message}`);
+        socket.emit('error:server', { message: err.message });
+      }
     });
 
     // Handle PM2 standardization
     socket.on('standardize:start', async (rawData) => {
+      try {
       const data = validateSocketData(standardizeStartSchema, rawData, socket, 'standardize:start');
       if (!data) return;
       const { repoPath, providerId } = data;
@@ -142,6 +148,10 @@ export function initSocket(io) {
       });
 
       console.log(`✅ Standardization complete: ${result.filesModified.length} files modified`);
+      } catch (err) {
+        console.error(`❌ Socket handler error [standardize:start]: ${err.message}`);
+        socket.emit('error:server', { message: err.message });
+      }
     });
 
     // Handle log streaming requests
@@ -280,116 +290,131 @@ export function initSocket(io) {
 
     // Handle error recovery requests (can trigger auto-fix agents)
     socket.on('error:recover', async (rawData) => {
-      const data = validateSocketData(errorRecoverSchema, rawData, socket, 'error:recover');
-      if (!data) return;
-      const { code, context } = data;
-      console.log(`🔧 Error recovery requested: ${code}`);
+      try {
+        const data = validateSocketData(errorRecoverSchema, rawData, socket, 'error:recover');
+        if (!data) return;
+        const { code, context } = data;
+        console.log(`🔧 Error recovery requested: ${code}`);
 
-      // Create auto-fix task
-      const task = await handleErrorRecovery(code, context);
+        // Create auto-fix task
+        const task = await handleErrorRecovery(code, context);
 
-      // Broadcast recovery task created
-      io.emit('error:recover:requested', {
-        code,
-        context,
-        taskId: task.id,
-        timestamp: Date.now()
-      });
+        // Broadcast recovery task created
+        io.emit('error:recover:requested', {
+          code,
+          context,
+          taskId: task.id,
+          timestamp: Date.now()
+        });
+      } catch (err) {
+        console.error(`❌ Socket handler error [error:recover]: ${err.message}`);
+        socket.emit('error:server', { message: err.message });
+      }
     });
 
     // App update handler — streams progress via socket
     socket.on('app:update', async (rawData) => {
-      const data = validateSocketData(appUpdateSchema, rawData, socket, 'app:update');
-      if (!data) return;
+      try {
+        const data = validateSocketData(appUpdateSchema, rawData, socket, 'app:update');
+        if (!data) return;
 
-      const app = await appsService.getAppById(data.appId);
-      if (!app) {
-        socket.emit('app:update:error', { message: 'App not found' });
-        return;
-      }
+        const app = await appsService.getAppById(data.appId);
+        if (!app) {
+          socket.emit('app:update:error', { message: 'App not found' });
+          return;
+        }
 
-      console.log(`⬇️ Socket update started for ${app.name}`);
-      const emit = (step, status, message) => {
-        socket.emit('app:update:step', { step, status, message, timestamp: Date.now() });
-      };
+        console.log(`⬇️ Socket update started for ${app.name}`);
+        const emit = (step, status, message) => {
+          socket.emit('app:update:step', { step, status, message, timestamp: Date.now() });
+        };
 
-      const result = await appUpdater.updateApp(app, emit).catch(err => {
-        socket.emit('app:update:error', { message: err.message });
-        return null;
-      });
+        const result = await appUpdater.updateApp(app, emit).catch(err => {
+          socket.emit('app:update:error', { message: err.message });
+          return null;
+        });
 
-      if (result) {
-        socket.emit('app:update:complete', { success: result.success, steps: result.steps });
-        console.log(`✅ Socket update complete for ${app.name}`);
+        if (result) {
+          socket.emit('app:update:complete', { success: result.success, steps: result.steps });
+          console.log(`✅ Socket update complete for ${app.name}`);
+        }
+      } catch (err) {
+        console.error(`❌ Socket handler error [app:update]: ${err.message}`);
+        socket.emit('error:server', { message: err.message });
       }
     });
 
     // App standardize handler — streams progress via socket
     socket.on('app:standardize', async (rawData) => {
-      const data = validateSocketData(appStandardizeSchema, rawData, socket, 'app:standardize');
-      if (!data) return;
+      try {
+        const data = validateSocketData(appStandardizeSchema, rawData, socket, 'app:standardize');
+        if (!data) return;
 
-      const app = await appsService.getAppById(data.appId);
-      if (!app) {
-        socket.emit('app:standardize:error', { message: 'App not found' });
-        return;
-      }
-
-      console.log(`🔧 Socket standardize started for ${app.name}`);
-      const emit = (step, status, message) => {
-        socket.emit('app:standardize:step', { step, status, message, timestamp: Date.now() });
-      };
-
-      // Step 1: Analyze
-      emit('analyze', 'running', 'Analyzing project configuration...');
-      const analysis = await pm2Standardizer.analyzeApp(app.repoPath)
-        .catch(err => ({ success: false, error: err.message }));
-
-      if (!analysis.success) {
-        emit('analyze', 'error', analysis.error);
-        socket.emit('app:standardize:error', { message: analysis.error });
-        return;
-      }
-      emit('analyze', 'done', `Found ${analysis.proposedChanges.processes?.length || 0} processes`);
-
-      // Step 2: Backup
-      emit('backup', 'running', 'Creating git backup...');
-      const backup = await pm2Standardizer.createGitBackup(app.repoPath)
-        .catch(err => ({ success: false, reason: err.message }));
-
-      if (backup.success) {
-        emit('backup', 'done', `Backup branch: ${backup.branch}`);
-      } else {
-        emit('backup', 'skipped', backup.reason || 'No git repository');
-      }
-
-      // Step 3: Apply
-      emit('apply', 'running', 'Writing ecosystem.config.cjs...');
-      const result = await pm2Standardizer.applyStandardization(app.repoPath, analysis)
-        .catch(err => ({ success: false, errors: [err.message] }));
-
-      if (result.errors?.length > 0) {
-        emit('apply', 'error', result.errors.join(', '));
-        socket.emit('app:standardize:error', { message: result.errors.join(', ') });
-        return;
-      }
-      emit('apply', 'done', `Modified ${result.filesModified.length} files`);
-
-      // Update app with new PM2 process names
-      if (analysis.proposedChanges?.processes) {
-        const pm2ProcessNames = analysis.proposedChanges.processes.map(p => p.name);
-        await appsService.updateApp(data.appId, { pm2ProcessNames });
-      }
-
-      socket.emit('app:standardize:complete', {
-        success: true,
-        result: {
-          backupBranch: result.backupBranch,
-          filesModified: result.filesModified,
-          processes: analysis.proposedChanges.processes
+        const app = await appsService.getAppById(data.appId);
+        if (!app) {
+          socket.emit('app:standardize:error', { message: 'App not found' });
+          return;
         }
-      });
-      console.log(`✅ Socket standardize complete for ${app.name}`);
+
+        console.log(`🔧 Socket standardize started for ${app.name}`);
+        const emit = (step, status, message) => {
+          socket.emit('app:standardize:step', { step, status, message, timestamp: Date.now() });
+        };
+
+        // Step 1: Analyze
+        emit('analyze', 'running', 'Analyzing project configuration...');
+        const analysis = await pm2Standardizer.analyzeApp(app.repoPath)
+          .catch(err => ({ success: false, error: err.message }));
+
+        if (!analysis.success) {
+          emit('analyze', 'error', analysis.error);
+          socket.emit('app:standardize:error', { message: analysis.error });
+          return;
+        }
+        emit('analyze', 'done', `Found ${analysis.proposedChanges.processes?.length || 0} processes`);
+
+        // Step 2: Backup
+        emit('backup', 'running', 'Creating git backup...');
+        const backup = await pm2Standardizer.createGitBackup(app.repoPath)
+          .catch(err => ({ success: false, reason: err.message }));
+
+        if (backup.success) {
+          emit('backup', 'done', `Backup branch: ${backup.branch}`);
+        } else {
+          emit('backup', 'skipped', backup.reason || 'No git repository');
+        }
+
+        // Step 3: Apply
+        emit('apply', 'running', 'Writing ecosystem.config.cjs...');
+        const result = await pm2Standardizer.applyStandardization(app.repoPath, analysis)
+          .catch(err => ({ success: false, errors: [err.message] }));
+
+        if (result.errors?.length > 0) {
+          emit('apply', 'error', result.errors.join(', '));
+          socket.emit('app:standardize:error', { message: result.errors.join(', ') });
+          return;
+        }
+        emit('apply', 'done', `Modified ${result.filesModified.length} files`);
+
+        // Update app with new PM2 process names
+        if (analysis.proposedChanges?.processes) {
+          const pm2ProcessNames = analysis.proposedChanges.processes.map(p => p.name);
+          await appsService.updateApp(data.appId, { pm2ProcessNames });
+        }
+
+        socket.emit('app:standardize:complete', {
+          success: true,
+          result: {
+            backupBranch: result.backupBranch,
+            filesModified: result.filesModified,
+            processes: analysis.proposedChanges.processes
+          }
+        });
+        console.log(`✅ Socket standardize complete for ${app.name}`);
+      } catch (err) {
+        console.error(`❌ Socket handler error [app:standardize]: ${err.message}`);
+        socket.emit('error:server', { message: err.message });
+      }
     });
 
     // App deploy handler — streams real-time output from deploy.sh
