+++ server/services/taskSchedule.js
@@ -2199,62 +2199,3 @@ function getTaskTypeDescription(taskType) {
   return descriptions[taskType] || taskType.replace(/-/g, ' ');
 }
 
-// ============================================================
-// Backward-compatible exports (delegate to unified functions)
-// ============================================================
-
-export async function getSelfImprovementInterval(taskType) {
-  return getTaskInterval(taskType);
-}
-
-export async function getAppImprovementInterval(taskType) {
-  return getTaskInterval(taskType);
-}
-
-export async function updateSelfImprovementInterval(taskType, settings) {
-  return updateTaskInterval(taskType, settings);
-}
-
-export async function updateAppImprovementInterval(taskType, settings) {
-  return updateTaskInterval(taskType, settings);
-}
-
-export async function shouldRunSelfImprovementTask(taskType) {
-  return shouldRunTask(taskType);
-}
-
-export async function shouldRunAppImprovementTask(taskType, appId) {
-  return shouldRunTask(taskType, appId);
-}
-
-export async function getDueSelfImprovementTasks() {
-  return getDueTasks();
-}
-
-export async function getDueAppImprovementTasks(appId) {
-  return getDueTasks(appId);
-}
-
-export async function getNextSelfImprovementTaskType(lastType) {
-  return getNextTaskType(null, lastType);
-}
-
-export async function getNextAppImprovementTaskType(appId, lastType) {
-  return getNextTaskType(appId, lastType);
-}
-
-export function getDefaultSelfImprovementPrompt(taskType) {
-  return getDefaultPrompt(taskType);
-}
-
-export function getDefaultAppImprovementPrompt(taskType) {
-  return getDefaultPrompt(taskType);
-}
-
-export async function getSelfImprovementPrompt(taskType) {
-  return getTaskPrompt(taskType);
-}
-
-export async function getAppImprovementPrompt(taskType) {
-  return getTaskPrompt(taskType);
-}
